feature: handle Numpy type deprecation & backwards compatibility
Closes #7

Detects Numpy version on the fly and sets the types for `float` and `complex` conditional on whether `np.__version__ < "2.0.0"`.

Includes a warning advising users to consider updating them if using an old version of Numpy.

No idea if this is a sensible way of approaching such issues so happy to be advised/guided on better ways of handling such issues.